+++ client/src/pages/RegisterPage/styles.ts
@@ -2,7 +2,6 @@ import { css } from '@emotion/react';
 
 import { COLORS } from 'styles/colors';
 import { COMMON_SIZE, FONT_SIZE } from 'styles/sizes';
-import { RESULT } from 'utils/constants';
 
 export const registerPageHeaderStyle = css({
   fontSize: FONT_SIZE.LARGE,
@@ -25,35 +24,6 @@ export const inputWrapperStyle = css({
   alignItems: 'center',
   justifyContent: 'center',
   position: 'relative',
-});
-
-export const dropdownContainerStyle = css({
-  width: '50%',
-  border: `1px solid ${COLORS.PRIMARY_1}`,
-  display: 'flex',
-  flexDirection: 'row',
-  justifyContent: 'space-between',
-  alignItems: 'center',
-  padding: 0,
-  paddingLeft: 10,
-
-  ' span': {
-    fontSize: FONT_SIZE.MEDIUM,
-    color: COLORS.TEXT_1,
-    marginRight: 20,
-  },
-});
-
-export const inputButtonArrowStyle = css({
-  backgroundColor: COLORS.PRIMARY_1,
-  padding: `0 10px`,
-  display: 'flex',
-  justifyItems: 'center',
-  alignItems: 'center',
-
-  ' svg': {
-    width: 20,
-  },
 });
 
 export const registerInputArrowButtonStyle = css({
